+++ mrb_doc/models/texture2d.rb
@@ -26,4 +26,25 @@ class Texture2D
     # src/mruby_integration/models/texture2d.cpp
     nil
   end
+
+  # Sets the filtering for the {Texture2D}.
+  #
+  # @example Basic usage
+  #   big_texture = Texture2D.new("/assets/my_cool_background.png")
+  #
+  #   big_texture.filter = Texture2D::BILINEAR
+  #
+  #   big_texture.unload
+  #
+  # @param val [Integer] What sort of filtering to apply, valid options are:
+  #   {Texture2D::NO_FILTER}, {Texture2D::BILINEAR}, {Texture2D::TRILINEAR},
+  #   {Texture2D::ANISOTROPIC_4X}, {Texture2D::ANISOTROPIC_8X}, or
+  #   {Texture2D::ANISOTROPIC_16X}.
+  # @return [nil]
+  # @raise [ArgumentError]
+  def filter=(val)
+    # mrb_Texture2D_filter_equals
+    # src/mruby_integration/models/texture2d.cpp
+    nil
+  end
 end